backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "A FastAPI backend for my chatbot"
authors = ["Ojas Kapre <ojasskapre@gmail.com>", "Sarvesh Kapre <sarveshkapre@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pydantic = {extras = ["email"], version = "^2.11.4"}
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
langchain = "^0.2.14"
openai = "^1.41.0"
httpx = "^0.27.0"
langchain-openai = "^0.1.22"
langchain-community = "^0.2.12"
python-dotenv = "^1.0.1"
supabase = "^2.7.2"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.9.1"
sentry-sdk = "^2.27.0"
colorlog = "^6.9.0"
python-multipart = "^0.0.20"

[tool.poetry.group.dev.dependencies]
fastapi = "^0.112.1"
uvicorn = "^0.30.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
